tests: separate tests
